flormar (2)/flormar/requirements.txt
asn1crypto==0.23.0
astroid==1.6.1
backcall==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
colorama==0.3.9
conda==4.3.31
cryptography==2.1.4
data==0.4
decorator==4.3.0
Django==1.11.10
django-debug-toolbar==1.9.1
django-ipware==2.0.1
djangorestframework==3.7.7
funcsigs==1.0.2
geoip2==2.7.0
google-api-python-client==1.6.5
googletrans==2.3.0
gspread==3.0.0
httplib2==0.11.1
idna==2.6
image==1.5.24
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
lazy-object-proxy==1.3.1
lxml==4.2.1
maxminddb==1.3.0
mccabe==0.6.1
menuinst==1.4.10
oauth2client==4.1.2
parso==0.2.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
py-translate==1.0.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycosat==0.6.3
pycparser==2.18
PyDrive==1.3.1
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.2
pyOpenSSL==17.5.0
PySocks==1.6.7
pytz==2018.3
pywin32==221
PyYAML==3.12
requests==2.18.4
retrying==1.3.3
rsa==3.4.2
ruamel-yaml==0.11.14
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
translation==1.0.5
twilio==6.13.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.10.11